fix: circular dependency in plugins test
