+++ src/app/components/Button/index.ts
@@ -1,1 +1,2 @@
 export * from "./Button";
+export * from "./OriginalButton";
